+++ components/31-molecules/breadcrumbs/breadcrumbs.bindings.js
@@ -1,0 +1,14 @@
+'use strict';
+
+(function () {
+
+  if (!Breadcrumbs) { // eslint-disable-line no-undef
+    return;
+  }
+
+  const selected = document.querySelectorAll('.breadcrumb');
+  for (let i = selected.length; i--;) {
+    new Breadcrumbs(selected[i]); // eslint-disable-line no-undef
+  }
+
+})();