+++ CHANGELOG.md
@@ -11,6 +11,7 @@ NOTE: Refer to upcoming changes in our README.md under "Roadmap"
 * Changed paddings for box with spot image at the top.
 * Renamed `document-box` to `documents`.
   > :warning: The `document-box` CSS class is deprecated as from now.
+* Make breadcrumbs collapsible.
 
 ### Fixed